+++ CHANGELOG.md
@@ -11,6 +11,27 @@
 ### Features
 
 * 添加登录页&完善全局样式文件 ([31776ea](https://github.com/Shideshanxx/vue-tsx-admin/commit/31776ea59852c7810fa80b6cf0b73a189d959deb))
+* 添加search-view页面&view-router动画及tags路由缓存修复 ([b09d98f](https://github.com/Shideshanxx/vue-tsx-admin/commit/b09d98f26481849d1cd9d0731634f67a4eb2d422))
+* 添加tags标签栏 ([61667e7](https://github.com/Shideshanxx/vue-tsx-admin/commit/61667e78cf4ff65dc0708281e8ef40a4c9f162a4))
+* 增加app预览页&修改app-icon&增加顶部nav导航栏 ([59fe45c](https://github.com/Shideshanxx/vue-tsx-admin/commit/59fe45cb9017f303fcd4a859c45bca77c9e39f00))
+* commitLint ([155088f](https://github.com/Shideshanxx/vue-tsx-admin/commit/155088f4a406129cf347525e2aef53184569bf9e))
+* prettierignore&layout/index ([82aa6ea](https://github.com/Shideshanxx/vue-tsx-admin/commit/82aa6ea6c1f9ea3e49aad3ad97bdf16f22bf72a0))
+
+
+
+# 0.0.0 (2021-11-29)
+
+
+### Bug Fixes
+
+* 修复侧边栏目录 ([c542982](https://github.com/Shideshanxx/vue-tsx-admin/commit/c5429820086a86755a44de827ce8a8fee89a2584))
+* 修复i18n无法找到messages的问题 ([802444b](https://github.com/Shideshanxx/vue-tsx-admin/commit/802444b0824bfc13f1020b08a7b170bf148d04ca))
+* lint-stage&changelog ([6e41fbd](https://github.com/Shideshanxx/vue-tsx-admin/commit/6e41fbd991f8a87fd0db9104d014383e8a897e2a))
+
+
+### Features
+
+* 添加登录页&完善全局样式文件 ([31776ea](https://github.com/Shideshanxx/vue-tsx-admin/commit/31776ea59852c7810fa80b6cf0b73a189d959deb))
 * 添加tags标签栏 ([61667e7](https://github.com/Shideshanxx/vue-tsx-admin/commit/61667e78cf4ff65dc0708281e8ef40a4c9f162a4))
 * 增加app预览页&修改app-icon&增加顶部nav导航栏 ([59fe45c](https://github.com/Shideshanxx/vue-tsx-admin/commit/59fe45cb9017f303fcd4a859c45bca77c9e39f00))
 * commitLint ([155088f](https://github.com/Shideshanxx/vue-tsx-admin/commit/155088f4a406129cf347525e2aef53184569bf9e))